pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-fixer"
version = "0.1.0"
description = "This program allows to clean image corrupted."
readme = "README.md"
authors = [
    { name = "Arnold Mokira", email = "dr.mokira@gmail.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["image-cleaning"]
dependencies = [
    "requests>=2.25.1",
    "python-dotenv==1.0.0",
    "Pillow",
]

# [project.urls]
# "Homepage" = "https://github.com/your-username/my-python-package"
# "Bug Tracker" = "https://github.com/your-username/my-python-package/issues"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
imfix = "image_fixer.__main__:main"

